Isolate SendToAnonExportDeleteList dropdown
