+++ FrontEnd/src/components/OrthancContent/ContentRootPanel.js
@@ -1,13 +1,12 @@
 import React, {Component, createRef} from 'react'
 import SearchForm from './SearchForm'
+import SendTo from '../CommonComponents/RessourcesDisplay/SendToAnonExportDeleteDropdown'
 import apis from '../../services/apis'
-
-import Dropdown from 'react-bootstrap/Dropdown'
 
 import TableSeriesFillFromParent from '../CommonComponents/RessourcesDisplay/TableSeriesFillFromParent'
 import TablePatientsWithNestedStudies from '../CommonComponents/RessourcesDisplay/TablePatientsWithNestedStudies'
 
-import {studyArrayToPatientArray, treeToStudyArray} from '../../tools/processResponse'
+import {studyArrayToPatientArray} from '../../tools/processResponse'
 
 import {connect} from 'react-redux'
 import {addStudiesToDeleteList} from '../../actions/DeleteList'
@@ -76,55 +75,6 @@ class ContentRootPanel extends Component {
         this.setState({currentSelectedStudyId: ''})
     }
 
-    /**
-     * return all study details
-     * of selected items
-     */
-    getStudySelectedDetails = () => {
-        let selectedIds = this.child.current.getSelectedRessources()
-
-        let studiesOfSelectedPatients = []
-
-        //Add all studies of selected patient
-        selectedIds.selectedPatients.forEach(orthancPatientId => {
-            //loop the redux and add all studies that had one of the selected patient ID
-            let studyArray = this.props.orthancContent.filter(study => {
-                if (study.ParentPatient === orthancPatientId) return true
-                else return false
-            })
-            //Add to the global list of selected studies
-            studiesOfSelectedPatients.push(...studyArray)
-        })
-
-        //add selected level studies
-        selectedIds.selectedStudies.forEach(element => {
-            this.props.orthancContent.forEach(study => {
-                if (element === study.ID)
-                    studiesOfSelectedPatients.push(study)
-            });
-        });
-
-        //Get only unique study ids
-        let uniqueSelectedOrthancStudyId = [...new Set(studiesOfSelectedPatients)];
-
-        return uniqueSelectedOrthancStudyId
-    }
-
-    sendToDeleteList = () => {
-        this.props.addStudiesToDeleteList(this.getStudySelectedDetails())
-    }
-
-    sendToAnonList = () => {
-        this.props.addStudiesToAnonList(this.getStudySelectedDetails())
-    }
-
-    sendToExportList = async () => {
-        //Get selected studies array
-        let selectedStudiesArray = treeToStudyArray(this.getStudySelectedDetails())
-        //Send it to redux
-        this.props.addStudiesToExportList(selectedStudiesArray)
-    }
-
     rowEventsStudies = {
         onClick: (e, row) => {
             this.setState({currentSelectedStudyId: row.StudyOrthancID})
@@ -141,10 +91,6 @@ class ContentRootPanel extends Component {
         return style;
     }
 
-    handleClick = (e) => {
-        e.stopPropagation()
-    }
-
     render = () => {
         return (
             <div className='jumbotron'>
@@ -152,21 +98,11 @@ class ContentRootPanel extends Component {
                 <div className='row'>
                     <div className='col-sm'>
                         <div className={'d-flex flex-row justify-content-between'}>
-                            <LabelDropdown selectedStudiesGetter={this.getStudySelectedDetails}/>
-                            <Dropdown onClick={this.handleClick}>
-                                <Dropdown.Toggle variant="warning" id="dropdown-basic">
-                                    Send To
-                                </Dropdown.Toggle>
-
-                                <Dropdown.Menu>
-                                    <Dropdown.Item className='bg-primary' onClick={this.sendToExportList}>Export
-                                        List</Dropdown.Item>
-                                    <Dropdown.Item className='bg-info' onClick={this.sendToAnonList}>Anonymize
-                                        List</Dropdown.Item>
-                                    <Dropdown.Item className='bg-danger' onClick={this.sendToDeleteList}>Delete
-                                        List</Dropdown.Item>
-                                </Dropdown.Menu>
-                            </Dropdown>
+                            <LabelDropdown selectedStudiesGetter={this.getStudySelectedDetails}/>                            
+                            <SendTo 
+                                studies={this.child.current===null ? [] : this.child.current.getSelectedRessources().selectedStudies} 
+                                patients={this.child.current===null ? [] : this.child.current.getSelectedRessources().selectedPatients}
+                            />
                         </div>
